Add Vagrantfile to allow easy testing.
The "vagrant" directory is populated with the Vagrantfile and
all other files required for provisioning a virtual machine for
easy testing. This virtual machine contains:

- Postfix and Dovecot installed and configured
- Golang installed
- Mutt installed and configured
- Everything from the "mailutils" package installed

Postfix and Dovecot are configured to function locally only. Mutt
is already configured to connect to Dovecot through IMAP.

Two accounts are already created in the configuration files. These
accounts (with passwords) are:

- mainaccount@mail.incognitotest:incognitomail
- webmaster@mail.incognitotest:incognitomail